add stat method for ssh validation, improve error handling for FS operations

+++ filediff-node-module/src/main/java/org/sudu/experiments/js/node/JsSftpClient.java
@@ -18,7 +18,7 @@ public interface JsSftpClient extends JSObject {
     @JSProperty
     int getMode();
     @JSProperty
-    int getSize();
+    double getSize();
     @JSProperty
     int getAtime();
     @JSProperty
